\name{do.states}
\alias{do.states}
\docType{data}
\title{do.states}
\description{
  The 36 unphased genotype states for the DO on the autosomes and X chromosome. Also the 8 DO founder letter codes.
}
\usage{do.states}
\format{
  A list frame with 3 elements.
  \describe{
    \item{\code{auto}}{Character vector with two letter codes for each of the possible DO genotype states.}
    \item{\code{X}}{List with two elements, FALSE and M, containing the genotype codes for the X chromosome.}
    \item{\code{founders}}{Character vector containing the founder letter codes.}
  }
}
\details{
  This contains the letter codes for each of the 36 unphased genotype states in the DO. It also contains the founder letters.
}
\examples{
  do.states
}
\keyword{datasets}
