\name{example.genes}
\alias{example.genes}
\docType{data}
\title{example.genes}
\description{
A set of genes from the Mouse Genome Informatics that are used in the documentation examples. (http://www.sanger.ac.uk/resources/mouse/genomes/) From Chr 7: 103 - 105 Mb on NCBI Build 38.
}
\source{
  http://informatics.jax.org/
}
\examples{
data(example.genes)
}
\keyword{datasets}
