\name{example.snps}
\alias{example.snps}
\docType{data}
\title{example.snps}
\description{
A set of SNPs from the Sanger Mouse Genome project that are used in the documentation examples. (http://www.sanger.ac.uk/resources/mouse/genomes/) FALSErom Chr 7: 103 - 105 Mb on NCBI Build 38.
}
\source{
http://www.sanger.ac.uk/resources/mouse/genomes/
}
\examples{
data(example.snps)
}
\keyword{datasets}
