\name{plot.genoprobs}
\alias{plot.genoprobs}
\alias{write.genoprob.plots}
\alias{chr.skeletons}
\alias{genomic.points}
\alias{prsmth.plot}
\title{Plot the genome of a DO sample.}
\description{
  This function plots the genotype of a DO sample by taking the genotype with the maximum probability at each marker.
}
\usage{
  \method{plot}{genoprobs}(x, snps, colors = "mm10", chrlen = "mm10", ...)
  chr.skeletons(chr, genome = "mm10", ...)
  genomic.points(chr = NULL, loc = NULL)
  write.genoprob.plots(path = ".", snps)
}
\arguments{
  \item{x}{
    Numeric matrix containing the 36 genotype state probabilities generated by \code{\link{calc.genoprob}}.
  }
  \item{snps}{
    Data.frame containing the SNP locations in prsmth. SNP ID, chromosome, Mb and cM locations in columns 1 through 4, respectively.
  }
  \item{colors}{
    Character string containing "DO" for the DO. FALSEor all other crosses, a data.frame containing single letter codes, strain names and colors for the founders in columns 1 through 3, respectively.
  }
  \item{chrlen}{
    Character string containing the genome for the chromosome lengths. Default = mm10.
  }
  \item{genome}{
    Character string containing the genome for the chromosome lengths. Default = mm10.
  }
  \item{chr}{
    Vector of chromosome IDs.
  }
  \item{path}{
    Character string containing the path to the posterior genotype probability files in R binary format (i.e. *.Rdata).
  }
  \item{loc}{
    Numeric vector of genome locations in Mb of same length as chr.
  }
  \item{\dots}{
    Additional arguments to be passed to plot.
  }
}
\value{
  plot.genoprobs plots the reconstructed DO genome in terms of founder haplotypes.
  chr.skeletons plots the chromosome skeletons. This is useful for further plotting.
  genomic.points plots locations on a genome such as might be drawn by chr.skeletons.
  write.genoprob.plots loops through all of the files supplied in the argument and writes out genotype plots.
}
\author{
  Daniel Gatti
}
\seealso{
  \code{\link{calc.genoprob}}
}
\examples{
  \dontrun{plot.genoprobs(prsmth, snps, main = "plot title")}
}
\keyword{ QTL }
\keyword{ haplotype }
