\name{get.do.states}
\alias{get.do.states}
\title{
  Get the 36 genotype states for the DO
}
\description{
  Get the 36 genotype states for the DO
}
\usage{
  get.do.states()
}
\value{
  Character vector containing the two letter genotype codes for the 36 genotype states.
}
\author{
  Daniel Gatti
}
\examples{
  get.do.states()
}
\keyword{ genotypes }
