\name{get.machine.precision}
\alias{get.machine.precision}
\title{
  Get the machine precsion
}
\description{
  Get the machine precision from .Machine$double.eps on a log10 scale.
}
\usage{
get.machine.precision()
}
\value{
  Numeric value containing the machine precision.
}
\author{
  Daniel Gatti
}
\examples{
  get.machine.precision()
}
\keyword{ precision }
