\name{internal_functions}
\alias{assign.autosome.femaleX.cases}
\alias{autosome.femaleX.trans.probs}
\alias{DO_autosome_recomb_freq}
\alias{DO_femaleX_recomb_freq}
\alias{DO_maleX_recomb_freq}
\alias{femaleX.trans.probs}
\alias{maleX.trans.probs}
\alias{emission_prob_from_r}
\alias{filter_smooth_allele}
\alias{filter_smooth_intensity}
\alias{initialize.hmm}
\alias{kinship_from_r}
\alias{strip.quality.columns}
\alias{tabulate.geno}
\alias{check.args}
\alias{check.args}
\alias{convert.pos.to.GRanges}
\alias{convert.pos.to.bp}
\alias{viterbi_from_r}
\title{
  Internal functions for DOQTL
}
\description{
  Not intended for external use.
}
\author{
  Daniel Gatti and Karl Broman
}
\keyword{internal}
