\name{kinship}
\alias{kinship}
\title{Create a kinship matrix based on allele sharing.}
\description{
  Read in the genotypes and produce a kinship matrix based on allele sharing.
}
\usage{
  kinship(genotype.file)
}
\arguments{
  \item{genotype.file}{
     Character, the full path to the genotype file.  This will ususally be the filtered and batch normalized X intensities as generated by \code{\link{filter.samples}}.
  }
}
\value{
  Numeric matrix, with rows and columns equal to the number of samples in the genotype file.
}
\author{
  Daniel Gatti
}
\note{
}
\seealso{
  \code{\link{extract.raw.data}}, \code{\link{filter.samples}}, \code{\link{batch.normalize}, \code{\link{calc.genoprob}}}
}
\examples{
  K = kinship("/demo/MUGA/genotypes.filt.txt")
}
\keyword{ MUGA }
\keyword{ kinship }

