\name{muga.snps.to.keep}
\alias{muga.snps.to.keep}
\docType{data}
\title{ SNPs to use for genotyping and mapping on the MUGA }
\description{
  SNPs to use for genotyping and mapping on the MUGA. These are a combination of one culled by UNC due to poor performance or uncertain location and ones culled by JAX based on low intensity values.
}
\usage{ muga.snps.to.keep }
\format{
  A character vector with MUGA SNP IDs to use when genotyping or mapping.
}
\examples{
  muga.snps.to.keep
}
\keyword{datasets}
