\name{plot.genotype.max}
\alias{plot.genotype.max}
\title{Plot the genome of a DO sample.}
\description{
  This function plots the genotype of a DO sample by taking the genotype with the maximum probability at each marker.
}
\usage{
  plot.genotype.max(prsmth, snps, colors = "DO", ...)
}
\arguments{
  \item{prsmth}{
    Numeric matrix containing the 36 genotype state probabilities generated by \code{\link{calc.genoprob}}.
  }
  \item{snps}{
    Data.frame containing the SNP locations in prsmth. SNP ID, chromosome, Mb and cM locations in columns 1 through 4, respectively.
  }
  \item{colors}{
    Character string containing "DO" for the DO. For all other crosses, a data.frame containing single letter codes, strain names and colors for the founders in columns 1 through 3, respectively.
  }
  \item{...}{
    Other options to be passed to plot().  Typically this would be xlim and would be used to limit the range to plot around the QTL.
  }
}
\value{
  No returned value.  Creates a plot of the reconstructed DO genome.
}
\author{
  Daniel Gatti
}
\seealso{
  \code{\link{calc.genoprob}}
}
\examples{
  \dontrun{plot.genotype.max(prsmth, snps, main = "plot title")}
}
\keyword{ QTL }
\keyword{ haplotype }
