\name{read.cross}
\alias{read.cross}
\title{Read in the data to genotype a DO cross.}
\description{
  This function reads in the X & Y intensities, sex and DO outbreeding generation for a DO population.
}
\usage{
  read.cross(x.file, y.file, sex.gen.file)
}
\arguments{
  \item{x.file}{
     Character, the full path to the X intenstity file.  This will ususally be the filtered and batch normalized X intensities as generated by \code{\link{filter.samples}} or \code{\link{batch.normalize}}.
  }
  \item{y.file}{
     Character, the full path to the Y intenstity file.  This will ususally be the filtered and batch normalized Y intensities as generated by \code{\link{filter.samples}} or \code{\link{batch.normalize}}.
  }
  \item{sex.gen.file}{
     Character, tab delimited file with 2 columns and row.names.  row.names must be the sample IDs and must match those in x.file and y.file.  Column1 must be the sex, must be called "sex" and must contain only "M", "F" or NA.  Column 2 must contain the outbreeding generation, must be called "gen" and must contain integers for each sample.
  }
}
\value{
  List, with three items, a matrix of X intensities, a matrix of Y intensities and a data.frame with sex and DO generation.
}
\author{
  Daniel Gatti
}
\note{
  The number of samples and SNPs read in, along with a sample count broken down by sex an DO generation is printed out.
  Future releases may include more error checking upon reading in the data.
}
\seealso{
  \code{\link{extract.raw.data}}, \code{\link{filter.samples}}, \code{\link{batch.normalize}, \code{\link{calc.genoprob}}}
}
\examples{
  data(snps)
  read.cross(path = "/demo/MUGA/", snps)
}
\keyword{ MUGA }
\keyword{ intensity }
\keyword{ normalization }
