\name{sex.founders}
\alias{sex.founders}
\docType{data}
\title{
  Data.frame with founder IDs and sex value.
}
\description{
  The sex of the founders and F1s.
}
\usage{data(sex.founders)}
\format{
  A data frame with 80 observations on the following 2 variables.
  \describe{
    \item{\code{strain}}{a factor with levels \code{AA} \code{AB} \code{AC} \code{AD} \code{AE} \code{AH} \code{BB} \code{BD} \code{BE} \code{BF} \code{BH} \code{CC} \code{CD} \code{CE} \code{CH} \code{DD} \code{DE} \code{DF} \code{DH} \code{EE} \code{EG} \code{FF} \code{GG} \code{GH} \code{HH}}
    \item{\code{sex}}{a factor with levels \code{F} \code{M}}
  }
}

\examples{
  data(sex.founders)
}
\keyword{datasets}
