\name{SNPmanipulation-internal}
\alias{check.args}
\alias{convert.pos.to.bp}
\alias{convert.pos.to.GRanges}
\alias{create.pos}
\title{Accessing, subsetting and plotting mouse SNPs.}
\description{Internal functions not called by the user.}
\author{Daniel Gatti: Dan.Gatti@jax.org}
\keyword{internal}

