\name{SNPmanipulation-package}
\alias{SNPmanipulation-package}
\alias{SNPmanipulation}
\docType{package}
\title{Accessing, subsetting and plotting mouse SNPs.}
\description{
There are several large mouse SNP data sets (Sanger, MouseHapMap, Perlegen, UNC).  This package allows users to access these data sets once they have been compiled into one file and indexed with Tabix (http://samtools.sourceforge.net/tabix.shtml). There are functions to retrieve SNPs, subset them for specific strains, obtain only SNPs that are ploymorphic for a subset of strains, plot SNPs and look for certain allele pattens in SNPs.
}
\details{
\tabular{ll}{
  Package: \tab SNPmanipulation\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Dependencies: R (>= 2.10), IRanges, GenomicRanges, Biostrings, Rsamtools
  Date: \tab 2012-01-16\cr
  License: \tab GPL-3\cr
}
}
\author{Daniel Gatti: Dan.Gatti@jax.org}
